1.12.2-2.1.8
Retweaked Config settings to promote more volcanos.
Reorganized config elements so that it's easier to find the section you
want to modify.
Added the ability to ignore tile entities thus allowing volcanos to
erupt in player and villager builds.
added a maxChance to rarity for ore generation in volcanic walls
changed the order of code so that volcanic walls will generate more
often
added a nodule generation chance config for controlling the rarity at
wich a stone block is spawned next to a generated ore to protect it from
in world smelting