1.12.2-2.7.14
Release Build

We can now prohibit lava interaction with mods that react badly to
having their blocks burned! For example Biomes O'Plenty.

There are 4 types of Volanos that can erupt now: small, medium, large,
and random.

Large is recommended.

I'm using a pattern when the source block's chance to erupt comes due.
Lava source blocks in the pattern are elected for eruption and will do
so. Large is a 3x3 pattern. The chance that all 9 blocks will have a
lava source block are not 100% and this gives us a change volcano shape.

For extreme hell where basically everything below y10 comes to the
surface, set volcano spawn chance to 90.