1.12.2-2.10.17
Pulled and merged TheGreatGeek's changes that allow you to set which
blocks will not smelt in the world.

Added a new option to allow you to specify other dimensions to allow the
lava dynamics to run in. Does not override the 3 default dimensions
boolean settings.

Change the noise parameter defaults to allow the volcanos to make a
better cone structure.

Changed the wall building algorithm to look for air in the cardinal
directions instead of the up direction to increase the chance that walls
form inside the lava plume.